pyproject.toml
[tool.poetry]
name = "repl_nix_Slopy-finalcode"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
flatten-json = "^0.1.13"
TagScriptEngine = "^2.3.0"
humanfriendly = "^10.0"
wavelink = "^1.3.2"
nextcord = "^2.2.0"
tinydb = "^4.7.0"
Pil-Lite = "^0.1.1"
Pillow = "^9.2.0"
helpers = "^0.2.0"
akinator = "^1.0.3"
utils = "^1.0.1"



[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
